requirements.txt
pyyaml
aiogram